Fix app icon generator

+++ apptools/image/image/svg2png.py
@@ -9,6 +9,6 @@ def svg2png(filecontent, scale, path, size=None):
         cairosvg.svg2png(bytestring=bytestring, write_to=path, scale=scale)
     else:
         try:
-            cairosvg.svg2png(bytestring=bytestring, write_to=path, scale=scale, parent_width=int(size.split("x")[0]), parent_height=int(size.split("x")[1]))
+            cairosvg.svg2png(bytestring=bytestring, write_to=path, scale=scale, parent_width=float(size.split("x")[0]), parent_height=float(size.split("x")[1]))
         except:
             print('svg2png failed')
